Carrot-Assistant/requirements.txt
haystack-ai==2.1.1
huggingface_hub==0.22.2
python-dotenv==1.0.1
llama-cpp-haystack==0.4.0
fastapi==0.110.3
uvicorn==0.29.0
torch==2.3.0
sse-starlette==2.1.2
psycopg2-binary==2.9.9
streamlit==1.36.0
rapidfuzz==3.5.2
sqlalchemy==2.0.30
openssl=3.1.4
sseclient-py==1.8.0